Big optimization: removed allocations and add opt-outs.
This is a fairly significant optimization changeset that removes
many of the unnecessary heap memory allocations that were present.
Benefiting from this does require the caller to pass adequate
preallocated memory.  This is optional, and the original interface
still works the way it did, and will necessarily allocate its
own memory on the GC heap (but still not as much as it did before
this commit).

There are also some ddoc comment fixups and related style
corrections in this commit.