Better input validation. Less purity.
The address parser now checks for its input address to overlap with its
output buffer, which can help with detecting memory corruption or
buffering mistakes.

Templated functions were marked as pure, but this is a difficult position
to take, because the caller can determine the purity based on which type
they pass into the template.  As such, the "pure" attribute was removed
from some AddressParseResult methods to allow a wider range of template
arguments (buffered writers).